+++ TrackedUltrasound/Content/Rendering/Shaders/VPRTBasicLightingVertexShader.hlsl
@@ -1,58 +1,58 @@
-Texture2D<float4> Texture : register(t0);
+Texture2D<min16float4> Texture : register(t0);
 sampler Sampler : register(s0);
 
 cbuffer Parameters : register(b0)
 {
-  float4 DiffuseColor             : packoffset(c0);
-  float3 EmissiveColor            : packoffset(c1);
-  float3 SpecularColor            : packoffset(c2);
-  float  SpecularPower            : packoffset(c2.w);
+  min16float4 DiffuseColor             : packoffset(c0);
+  min16float3 EmissiveColor            : packoffset(c1);
+  min16float3 SpecularColor            : packoffset(c2);
+  min16float  SpecularPower            : packoffset(c2.w);
 
-  float3 LightDirection[3]        : packoffset(c3);
-  float3 LightDiffuseColor[3]     : packoffset(c6);
-  float3 LightSpecularColor[3]    : packoffset(c9);
+  min16float3 LightDirection[3]        : packoffset(c3);
+  min16float3 LightDiffuseColor[3]     : packoffset(c6);
+  min16float3 LightSpecularColor[3]    : packoffset(c9);
 
-  float4x4 World                  : packoffset(c12);
-  float3x3 WorldInverseTranspose  : packoffset(c16);
+  min16float4x4 World                  : packoffset(c12);
+  min16float3x3 WorldInverseTranspose  : packoffset(c16);
 };
 
 // A constant buffer that stores each set of view and projection matrices in column-major format.
 cbuffer ViewProjectionConstantBuffer : register(b1)
 {
-  float4 EyePosition[2];
-  float4x4 ViewProjection[2];
+  min16float4 EyePosition[2];
+  min16float4x4 ViewProjection[2];
 };
 
 struct VertexShaderOutput
 {
-  float4 PositionPS : SV_Position;
-  float4 Diffuse    : COLOR0;
-  float4 Specular   : COLOR1;
+  min16float4 PositionPS : SV_Position;
+  min16float4 Diffuse    : COLOR0;
+  min16float4 Specular   : COLOR1;
   uint rtvId        : SV_RenderTargetArrayIndex; // SV_InstanceID % 2
 };
 
 struct VertexShaderInput
 {
-  float4 Position : SV_Position;
-  float3 Normal   : NORMAL0;
-  float4 Tangent  : TANGENT0;
-  float4 Color    : COLOR0;
-  float2 TexCoord : TEXCOORD0;
+  min16float4 Position : SV_Position;
+  min16float3 Normal   : NORMAL0;
+  min16float4 Tangent  : TANGENT0;
+  min16float4 Color    : COLOR0;
+  min16float2 TexCoord : TEXCOORD0;
   uint instId     : SV_InstanceID;
 };
 
 struct ColorPair
 {
-  float3 Diffuse;
-  float3 Specular;
+  min16float3 Diffuse;
+  min16float3 Specular;
 };
 
-ColorPair ComputeLights(float3 eyeVector, float3 worldNormal, uniform int numLights)
+ColorPair ComputeLights(min16float3 eyeVector, min16float3 worldNormal, uniform int numLights)
 {
-  float3x3 lightDirections = 0;
-  float3x3 lightDiffuse = 0;
-  float3x3 lightSpecular = 0;
-  float3x3 halfVectors = 0;
+  min16float3x3 lightDirections = 0;
+  min16float3x3 lightDiffuse = 0;
+  min16float3x3 lightSpecular = 0;
+  min16float3x3 halfVectors = 0;
 
   [unroll]
   for (int i = 0; i < numLights; i++)
@@ -64,13 +64,13 @@ ColorPair ComputeLights(float3 eyeVector, float3 worldNormal, uniform int numLig
     halfVectors[i] = normalize(eyeVector - lightDirections[i]);
   }
 
-  float3 dotL = mul(-lightDirections, worldNormal);
-  float3 dotH = mul(halfVectors, worldNormal);
+  min16float3 dotL = mul(-lightDirections, worldNormal);
+  min16float3 dotH = mul(halfVectors, worldNormal);
 
-  float3 zeroL = step(0, dotL);
+  min16float3 zeroL = step(0, dotL);
 
-  float3 diffuse = zeroL * dotL;
-  float3 specular = pow(max(dotH, 0) * zeroL, SpecularPower);
+  min16float3 diffuse = zeroL * dotL;
+  min16float3 specular = pow(max(dotH, 0) * zeroL, SpecularPower);
 
   ColorPair result;
 
@@ -86,17 +86,17 @@ VertexShaderOutput main(VertexShaderInput vin)
 
   int idx = vin.instId % 2;
 
-  float4 pos_ws = mul(vin.Position, World);
-  float3 eyeVector = normalize(EyePosition[idx].xyz - pos_ws.xyz);
-  float3 worldNormal = normalize(mul(vin.Normal, WorldInverseTranspose));
+  min16float4 pos_ws = mul(vin.Position, World);
+  min16float3 eyeVector = normalize(EyePosition[idx].xyz - pos_ws.xyz);
+  min16float3 worldNormal = normalize(mul(vin.Normal, WorldInverseTranspose));
 
   ColorPair lightResult = ComputeLights(eyeVector, worldNormal, 1);
 
-  float4x4 worldViewProj = mul(ViewProjection[idx], World);
+  min16float4x4 worldViewProj = mul(ViewProjection[idx], World);
 
   vout.PositionPS = mul(vin.Position, worldViewProj);
-  vout.Diffuse = float4(lightResult.Diffuse, DiffuseColor.a);
-  vout.Specular = float4(lightResult.Specular, 1);
+  vout.Diffuse = min16float4(lightResult.Diffuse, DiffuseColor.a);
+  vout.Specular = min16float4(lightResult.Specular, 1);
 
   vout.Diffuse *= vin.Color;
